Use Firebase Remote Config to check against minimum and latest app versions

+++ lib/features/farm_shop_manager/data/repository/farm_shop_manager_repository.dart
@@ -1,4 +1,3 @@
-import 'package:farmhub/core/auth/domain/i_auth_repository.dart';
 import 'package:farmhub/core/auth/global_auth_cubit/global_auth_cubit.dart';
 import 'package:farmhub/core/network/network_info.dart';
 import 'package:farmhub/features/farm_shop_manager/data/datasources/farm_shop_manager_local_datasource.dart';
@@ -9,6 +8,8 @@ import 'package:farmhub/core/typedefs/typedefs.dart';
 import 'package:farmhub/features/farm_shop_manager/domain/i_farm_shop_manager_repository.dart';
 import 'package:fpdart/fpdart.dart';
 
+import '../../../../core/app_version_helper/app_version_helper.dart';
+import '../../../../core/auth/data/repository/auth_repository.dart';
 import '../../../../core/errors/exceptions.dart';
 import '../../../../core/errors/failures.dart';
 import '../../../../core/util/app_const.dart';
@@ -32,6 +33,16 @@ class FarmShopManagerRepository implements IFarmShopManagerRepository {
   FutureEither<Farm> createFarm({required String farmName, required Address farmAddress}) async {
     if (await networkInfo.isConnected) {
       try {
+        // Check the user's app version
+        bool isAllowed = await AppVersionHelper.isAppVersionAllowed();
+        if (!isAllowed) {
+          return Left(AuthFailure(
+            code: ERR_APP_VERSION_NOT_SUPPORTED,
+            message: MESSAGE_APP_VERSION_NOT_SUPPORTED,
+            stackTrace: StackTrace.current,
+          ));
+        }
+
         final user = globalAuthCubit.state.farmhubUser!;
 
         final result = await remoteDatasource.createFarm(
@@ -67,6 +78,16 @@ class FarmShopManagerRepository implements IFarmShopManagerRepository {
   FutureEither<Shop> createShop({required String shopName, required Address shopAddress}) async {
     if (await networkInfo.isConnected) {
       try {
+        // Check the user's app version
+        bool isAllowed = await AppVersionHelper.isAppVersionAllowed();
+        if (!isAllowed) {
+          return Left(AuthFailure(
+            code: ERR_APP_VERSION_NOT_SUPPORTED,
+            message: MESSAGE_APP_VERSION_NOT_SUPPORTED,
+            stackTrace: StackTrace.current,
+          ));
+        }
+
         final user = globalAuthCubit.state.farmhubUser!;
 
         final result = await remoteDatasource.createShop(
@@ -96,6 +117,16 @@ class FarmShopManagerRepository implements IFarmShopManagerRepository {
   FutureEither<Unit> deleteFarm({required String farmId}) async {
     if (await networkInfo.isConnected) {
       try {
+        // Check the user's app version
+        bool isAllowed = await AppVersionHelper.isAppVersionAllowed();
+        if (!isAllowed) {
+          return Left(AuthFailure(
+            code: ERR_APP_VERSION_NOT_SUPPORTED,
+            message: MESSAGE_APP_VERSION_NOT_SUPPORTED,
+            stackTrace: StackTrace.current,
+          ));
+        }
+
         final user = globalAuthCubit.state.farmhubUser!;
 
         await remoteDatasource.deleteFarm(farmhubUser: user, farmId: farmId);
@@ -121,6 +152,16 @@ class FarmShopManagerRepository implements IFarmShopManagerRepository {
   FutureEither<Unit> deleteShop({required String shopId}) async {
     if (await networkInfo.isConnected) {
       try {
+        // Check the user's app version
+        bool isAllowed = await AppVersionHelper.isAppVersionAllowed();
+        if (!isAllowed) {
+          return Left(AuthFailure(
+            code: ERR_APP_VERSION_NOT_SUPPORTED,
+            message: MESSAGE_APP_VERSION_NOT_SUPPORTED,
+            stackTrace: StackTrace.current,
+          ));
+        }
+
         final user = globalAuthCubit.state.farmhubUser!;
 
         await remoteDatasource.deleteShop(farmhubUser: user, shopId: shopId);
@@ -146,6 +187,16 @@ class FarmShopManagerRepository implements IFarmShopManagerRepository {
   FutureEither<Unit> updateFarm({required Farm farm}) async {
     if (await networkInfo.isConnected) {
       try {
+        // Check the user's app version
+        bool isAllowed = await AppVersionHelper.isAppVersionAllowed();
+        if (!isAllowed) {
+          return Left(AuthFailure(
+            code: ERR_APP_VERSION_NOT_SUPPORTED,
+            message: MESSAGE_APP_VERSION_NOT_SUPPORTED,
+            stackTrace: StackTrace.current,
+          ));
+        }
+
         final user = globalAuthCubit.state.farmhubUser!;
 
         await remoteDatasource.updateFarm(farmhubUser: user, farm: farm);
@@ -171,6 +222,16 @@ class FarmShopManagerRepository implements IFarmShopManagerRepository {
   FutureEither<Unit> updateShop({required Shop shop}) async {
     if (await networkInfo.isConnected) {
       try {
+        // Check the user's app version
+        bool isAllowed = await AppVersionHelper.isAppVersionAllowed();
+        if (!isAllowed) {
+          return Left(AuthFailure(
+            code: ERR_APP_VERSION_NOT_SUPPORTED,
+            message: MESSAGE_APP_VERSION_NOT_SUPPORTED,
+            stackTrace: StackTrace.current,
+          ));
+        }
+
         final user = globalAuthCubit.state.farmhubUser!;
 
         await remoteDatasource.updateShop(farmhubUser: user, shop: shop);
